qtvcp -rename screenoptions
